update todo APIs and db schema

+++ controllers/todos.js
@@ -18,8 +18,7 @@ module.exports = {
         userId: req.session.userId,
         title: req.body.title,
         body: req.body.body,
-        startDate: req.body.startDate,
-        endDate: req.body.endDate,
+        time: req.body.time
       })
         .then((data) => res.status(201).send(data))
         .catch(() => res.status(400).send({ error: '400 Bad Request' }));
@@ -28,26 +27,25 @@ module.exports = {
     }
   },
   update: async (req, res) => {
-    if (req.session.userId) {
+    if (req.session.userId && req.body.id) {
       await Todos.update(
         {
           title: req.body.title,
           body: req.body.body,
-          startDate: req.body.startDate,
-          endDate: req.body.endDate,
+          time: req.body.time,
+          isDone: req.body.isDone
         },
         {
           where: {
             userId: req.session.userId,
-            id: req.params.id
+            id: req.body.id
           }
         }
       )
-
       await Todos.findOne({
         where: {
           userId: req.session.userId,
-          id: req.params.id
+          id: req.body.id
         }
       }).then((data) => res.status(200).send(data))
         .catch(() => res.status(404).send({ error: '404 Not Found' }));
@@ -56,21 +54,16 @@ module.exports = {
     }
   },
   delete: (req, res) => {
-    if (req.session.userId) {
+    if (req.session.userId && req.body.id) {
       Todos.destroy({
         where: {
           userId: req.session.userId,
-          id: req.params.id
+          id: req.body.id
         }
       })
         .then(() => res.status(200).send({ result: 'Success' }))
         .catch(() => res.status(400).send({ error: '400 Bad Request' }));
-    } else {
-      res.status(400).send({ error: '400 Bad Request' });
-    }
-  },
-  deleteAll: (req, res) => {
-    if (req.session.userId) {
+    } else if (req.session.userId) {
       Todos.destroy({
         where: {
           userId: req.session.userId
